+++ models/student.py
@@ -8,13 +8,6 @@ class Student:
         self.roll_no = roll_no
         self.marks = marks
 
-    def total_marks(self):
-        return sum(self.marks)
-
-    def percentage(self):
-        return self.total_marks() / len(self.marks)
-
-    # convert object to dictionary for JSON storage
     def to_dict(self):
         return {
             "student_id": self.student_id,